While on pokemon details site hides display only favs button as well as view options button
