+++ src/components/Header/Header.tsx
@@ -1,18 +1,11 @@
 import "./Header.scss";
 import { Button } from "../Button/Button.tsx";
 import { Link } from "react-router-dom";
-import { ViewOptionsButton } from "./HeaderButtons/ViewOptionsButton.tsx";
 import { DarkModeButton } from "./HeaderButtons/DarkModeButton.tsx";
 import { SearchBar } from "./SearchBar.tsx";
-import { FiltersContext } from "../../contexts/FiltersContextProvider.tsx";
-import { useContext } from "react";
+import { ShowViewOptionsAndFavsButtons } from "./ShowViewOptionsAndFavsButtons.tsx";
 
 export const Header = () => {
-  const { isOnlyFavs, setNewIsOnlyFavs } = useContext(FiltersContext);
-  const handleOnlyFavs = () => {
-    setNewIsOnlyFavs(!isOnlyFavs);
-  };
-
   return (
     <div className="flex-row space-between align-center max-width-1440 header">
       <Button
@@ -27,10 +20,7 @@ export const Header = () => {
       </Button>
       <SearchBar />
       <div className={"flex-row gap-18"}>
-        <ViewOptionsButton />
-        <Button className={"header__btn"} onClick={handleOnlyFavs}>
-          <i className="fa-solid fa-heart header__icon"></i>
-        </Button>
+        <ShowViewOptionsAndFavsButtons />
         <DarkModeButton />
       </div>
     </div>
